+++ src/methods/Method.hpp
@@ -23,7 +23,7 @@ public:
     void checkLockingBeforeReport(Alignment A);
     void checkNodeTypesBeforeReport(Alignment A);
 
-    unordered_map<ushort,ushort> getReverseMap(const unordered_map<ushort,ushort> reverse) const;
+    unordered_map<uint,uint> getReverseMap(const unordered_map<uint,uint> reverse) const;
 
 protected:
     Graph* G1;
